+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Hopully fixed shields going places in the Creative Menu
+
 ## 1.17.2
 * Added config option (under Mining) to enable Beeg Ore Veins
   * Now generate in biomes where more ores generate